typecast for warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@16570 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ tools/tankview/simpleSDLView.cpp
@@ -517,14 +517,14 @@ bool SimpleDisplay::update ( void )
       {
 	int x,y;
 	SDL_GetMouseState(&x,&y);
-	y = size[1]-y;
+	y = (int)size[1]-y;
 
    	mouseButton(event.button.button,x,y,event.type == SDL_MOUSEBUTTONDOWN);
       }
       break;
 
     case SDL_MOUSEMOTION:
-	mouseMoved(event.motion.x,size[1]-event.motion.y);
+	mouseMoved(event.motion.x,(int)size[1]-event.motion.y);
       break;
 
     }
